requirements.txt
tensorrt    # 8.2.1.5
onnx
onnxruntime
nvidia-pyindex
onnx-graphsurgeon
paddleslim
numpy
tqdm
typer
cuda-python
colored
rich

# TODO: add version

# install requirements in order:
# cat requirements.txt | xargs -n 1 -L 1 pip install